requirements.txt
numpy
imageio
torchvision
matplotlib
tqdm
pytorch_msssim
lpips
scipy
open3d
trimesh
